update_project

+++ frontend/src/pages/StudentsView/StudentsView.vue
@@ -7,23 +7,24 @@ import VModalWindow from '@/components/UI/VModalWindow/VModalWindow.vue'
 import VButton from '@/components/UI/VButton/VButton.vue'
 import VSelect from '@/components/UI/VSelect/VSelect.vue'
 import studentService from '@/API/studentService.js'
-import { useStudentStore } from '@/store/studentStore.js'
-import { ref } from 'vue'
-import { useGroupStore } from '@/store/groupStore.js'
+import {useStudentStore} from '@/store/studentStore.js'
+import {ref} from 'vue'
+import {useGroupStore} from '@/store/groupStore.js'
+import VInformationWindow from "@/components/VInformationWindow/VInformationWindow.vue";
 
 const studentStore = useStudentStore()
+await studentStore.getStudents()
+
 const groupStore = useGroupStore()
 
 const addStudentWindowOpened = ref(false)
 const editStudentWindowOpened = ref(false)
-
 const dataStudent = ref({
   fio: '',
   address: '',
   group: '',
   phone: '',
 })
-
 const dataStudentForEdit = ref({
   id: '',
   fio: '',
@@ -31,7 +32,6 @@ const dataStudentForEdit = ref({
   group: '',
   phone: '',
 })
-
 const tableHead = [
   {
     id: 1,
@@ -54,6 +54,12 @@ const tableHead = [
     name: 'телефон',
   },
 ]
+const students = ref(studentStore.students)
+const dataTableStudent = ref(students.value)
+const windowInfoText = ref('')
+const windowInfoVisible = ref(false)
+const dialogWindowVisible = ref(false)
+const idStudent = ref()
 
 const switchAddStudentWindow = () => {
   addStudentWindowOpened.value = !addStudentWindowOpened.value
@@ -75,47 +81,52 @@ const switchEditStudentWindow = (data) => {
 const addStudent = async () => {
   const reqAddStudent = await studentService.addStudent(dataStudent.value)
   if (!reqAddStudent.errno) {
-    studentStore.students.push({
-      id: reqAddStudent.insertId,
-      fio: dataStudent.value.fio,
-      address: dataStudent.value.address,
-      group: dataStudent.value.group,
-      phone: dataStudent.value.phone,
-    })
-  } else {
-    console.log(reqAddStudent)
+    dataTableStudent.value.push(
+        {
+          id: reqAddStudent.insertId,
+          fio: dataStudent.value.fio,
+          address: dataStudent.value.address,
+          group: dataStudent.value.group,
+          phone: dataStudent.value.phone,
+        })
+    switchAddStudentWindow()
+    openInfoWindow(`Студент #${dataStudent.value.fio} успешно добавлен!`)
   }
 }
 
 const editStudent = async () => {
   const reqEditStudent = studentService.editStudent(
-    dataStudentForEdit.value,
-    dataStudentForEdit.value.id
+      dataStudentForEdit.value,
+      dataStudentForEdit.value.id
   )
   if (!reqEditStudent.errno) {
-    const index = studentStore.students.findIndex(
-      (student) => student.id === dataStudentForEdit.value.id
+    const index = dataTableStudent.value.findIndex(
+        (student) => student.id === dataStudentForEdit.value.id
     )
-    studentStore.students[index].fio = dataStudentForEdit.value.fio
-    studentStore.students[index].address = dataStudentForEdit.value.address
-    studentStore.students[index].group = dataStudentForEdit.value.group
-    studentStore.students[index].phone = dataStudentForEdit.value.phone
-  } else {
-    console.log(reqEditStudent)
+    dataTableStudent.value[index].fio = dataStudentForEdit.value.fio
+    dataTableStudent.value[index].address = dataStudentForEdit.value.address
+    dataTableStudent.value[index].group = dataStudentForEdit.value.group
+    dataTableStudent.value[index].phone = dataStudentForEdit.value.phone
+    openInfoWindow('Изменения прошли успешно!')
+    switchEditStudentWindow()
   }
 }
 
-const deleteById = async (id) => {
-  const reqDeleteStudent = await studentService.deleteStudentById(id)
-  if (!reqDeleteStudent.errno) {
-    const index = studentStore.students.findIndex(
-      (student) => student.id === id
-    )
-    if (index !== -1) {
-      studentStore.students.splice(index, 1)
+const deleteById = async (action, id) => {
+  if (action === 'delete') {
+    const reqDeleteStudent = await studentService.deleteStudentById(id)
+    if (!reqDeleteStudent.errno) {
+      const index = dataTableStudent.value.findIndex(
+          (student) => student.id === id
+      )
+      if (index !== -1) {
+        dataTableStudent.value.splice(index, 1)
+      }
+      dialogWindowVisible.value = !dialogWindowVisible.value
+      openInfoWindow(`Студент #${id} успешно удален!`)
     }
-  } else {
-    console.log(reqDeleteStudent)
+  } else if (action === 'cancel') {
+    dialogWindowVisible.value = !dialogWindowVisible.value
   }
 }
 
@@ -124,29 +135,60 @@ const getNameGroup = (id) => {
   return groups.filter((group) => group.value === id)[0].title
 }
 
+const tableSearch = (data) => {
+  const input = data.inp
+  const select = data.select
+
+  if (input && select) {
+    dataTableStudent.value = students.value.filter((obj) =>
+        obj.fio.includes(data.inp) && obj.group === data.select)
+  } else if (input) {
+    dataTableStudent.value = students.value.filter((obj) =>
+        obj.fio.includes(input))
+  } else if (select) {
+    dataTableStudent.value = students.value.filter((obj) =>
+        obj.group === select)
+  } else {
+    dataTableStudent.value = students.value
+  }
+}
+
+const openInfoWindow = (text) => {
+  windowInfoText.value = text
+  windowInfoVisible.value = true
+  setTimeout(() => {
+    windowInfoVisible.value = false
+  }, 2000)
+}
+
+const openDialogWindow = (id) => {
+  idStudent.value = id
+  dialogWindowVisible.value = !dialogWindowVisible.value
+}
+
 await groupStore.getGroups()
-await studentStore.getStudents()
 </script>
 
 <template>
   <div class="students">
-    <VSidebar class="" />
+    <VSidebar/>
     <div class="students-body">
-      <VHeader />
+      <VHeader/>
       <h2 class="students-title">Студенты</h2>
       <VTable
-        @switchAddWindow="switchAddStudentWindow"
-        @deleteById="deleteById"
-        :data-table="studentStore.students"
-        :options="groupStore.groups"
-        :tableHead="tableHead"
-        title="Данные студентов"
+          @switchAddWindow="switchAddStudentWindow"
+          @deleteById="deleteById"
+          @search-by-inp-and-select="tableSearch"
+          :options="groupStore.groups"
+          :tableHead="tableHead"
+          title="Данные студентов"
+          placeholder-search="Поиск по ФИО"
       >
         <transition-group name="list" tag="p">
           <tr
-            class="table-body-row"
-            v-for="data in studentStore.students"
-            :key="data.id"
+              class="table-body-row"
+              v-for="data in dataTableStudent"
+              :key="data.id"
           >
             <td>{{ data.id }}</td>
             <td>{{ data.fio }}</td>
@@ -154,14 +196,14 @@ await studentStore.getStudents()
             <td>{{ getNameGroup(data.group) }}</td>
             <td>{{ data.phone }}</td>
             <td>
-              <VButton @click="deleteById(data.id)" size="icon" variant="text">
+              <VButton @click="openDialogWindow(data.id)" size="icon" variant="text">
                 <v-icon color="red">mdi-delete</v-icon>
               </VButton>
               <VButton
-                @click="switchEditStudentWindow(data)"
-                size="icon"
-                variant="text"
-                icon="mid-pencil"
+                  @click="switchEditStudentWindow(data)"
+                  size="icon"
+                  variant="text"
+                  icon="mid-pencil"
               >
                 <v-icon color="secondary">mdi-pencil</v-icon>
               </VButton>
@@ -175,39 +217,39 @@ await studentStore.getStudents()
         <h3 class="addStudent-title">Добавление студента</h3>
         <div class="addStudent-form">
           <VInput
-            v-model="dataStudent.fio"
-            size="fullWidth"
-            placeholder="ФИО"
+              v-model="dataStudent.fio"
+              size="fullWidth"
+              placeholder="ФИО"
           ></VInput>
           <VInput
-            v-model="dataStudent.address"
-            size="fullWidth"
-            placeholder="Адрес"
+              v-model="dataStudent.address"
+              size="fullWidth"
+              placeholder="Адрес"
           ></VInput>
           <VSelect
-            v-model="dataStudent.group"
-            size="fullWidth"
-            :options="groupStore.groups"
-            disabled-option="Группа"
+              v-model="dataStudent.group"
+              size="fullWidth"
+              :options="groupStore.groups"
+              disabled-option="Группа"
           ></VSelect>
           <VInput
-            v-model="dataStudent.phone"
-            size="fullWidth"
-            placeholder="Телефон"
+              v-model="dataStudent.phone"
+              size="fullWidth"
+              placeholder="Телефон"
           ></VInput>
           <div class="addStudent-actions">
             <v-button
-              @click="switchAddStudentWindow()"
-              size="small"
-              color="secondary"
-              >закрыть
+                @click="switchAddStudentWindow()"
+                size="small"
+                color="secondary"
+            >закрыть
             </v-button>
             <v-button
-              @click="addStudent()"
-              class="button-add"
-              size="small"
-              color="primary"
-              >добавить
+                @click="addStudent()"
+                class="button-add"
+                size="small"
+                color="primary"
+            >добавить
             </v-button>
           </div>
         </div>
@@ -221,41 +263,57 @@ await studentStore.getStudents()
         </h3>
         <div class="addStudent-form">
           <VInput
-            v-model="dataStudentForEdit.fio"
-            size="fullWidth"
-            placeholder="ФИО"
+              v-model="dataStudentForEdit.fio"
+              size="fullWidth"
+              placeholder="ФИО"
           ></VInput>
           <VInput
-            v-model="dataStudentForEdit.address"
-            size="fullWidth"
-            placeholder="Адрес"
+              v-model="dataStudentForEdit.address"
+              size="fullWidth"
+              placeholder="Адрес"
           ></VInput>
           <VSelect
-            v-model="dataStudentForEdit.group"
-            size="fullWidth"
-            :options="groupStore.groups"
-            disabled-option="Группа"
+              v-model="dataStudentForEdit.group"
+              size="fullWidth"
+              :options="groupStore.groups"
+              disabled-option="Группа"
           ></VSelect>
           <VInput
-            v-model="dataStudentForEdit.phone"
-            size="fullWidth"
-            placeholder="Телефон"
+              v-model="dataStudentForEdit.phone"
+              size="fullWidth"
+              placeholder="Телефон"
           ></VInput>
           <div class="addStudent-actions">
             <v-button
-              @click="switchEditStudentWindow()"
-              size="small"
-              color="secondary"
-              >закрыть
+                @click="switchEditStudentWindow()"
+                size="small"
+                color="secondary"
+            >закрыть
             </v-button>
             <v-button
-              @click="editStudent()"
-              class="button-add"
-              size="small"
-              color="primary"
-              >готово
+                @click="editStudent()"
+                class="button-add"
+                size="small"
+                color="primary"
+            >готово
             </v-button>
           </div>
+        </div>
+      </div>
+    </VModalWindow>
+    <VInformationWindow :visible="windowInfoVisible">
+      {{ windowInfoText }}
+    </VInformationWindow>
+    <VModalWindow :is-open="dialogWindowVisible">
+      <div class="groups_dialogDelete">
+        <h3 class="dialogDelete-title">Вы уверены, что хотите удалить студента #{{ idStudent }}?</h3>
+        <div class="dialogDelete-actions">
+          <VButton @click="deleteById('cancel', idStudent)" class="dialogDelete-cancel" size="small" color="primary"
+          >нет
+          </VButton>
+          <VButton @click="deleteById('delete', idStudent)" class="dialogDelete-accept" size="small" color="secondary"
+          >да
+          </VButton>
         </div>
       </div>
     </VModalWindow>
